adds a reader for compliance data types. moves the list compliance datatypes type definitions to the list directory.
